// client-new/src/generated/schema-metadata.generated.ts
export type ColumnMetadata = {
  schema: string;
  table: string;
  column: string;
  baseType: string;
  isArray: boolean;
  isNullable: boolean;
  enumRef: { schema: string; name: string } | null;
  references: Array<{ relation: string; referencedColumns: string[]; isOneToOne: boolean; schema?: string }>;
};